m3fs: implement session closing properly.
First, I reduced the storing of the FileSessions in OpenFile to a simple
reference count. All we need to know there is whether there are other
sessions still using the OpenFile; we don't care which sessions that
are.

Second, the MetaSession does no longer store FileSessions either, but
just their ids. For that reason, we don't need RefCells for
FileSessions anymore. Another positive side effect is that we can now
implement the cleanup for FileSession in drop since we don't have
references to FileSession elsewhere that would prevent drop from being
called.

Third, meta sessions now actually remove the file sessions that were
created by them. And they also prevent that more files are opened than
allowed.

So, in summary: both FileSession and MetaSession is removed and dropped
in the request handler on close. For the MetaSession, the request
handler also removes the contained FileSessions. For the FileSession,
the request handler removes it from the MetaSession. In general, the
FileSession cleans up everything on drop. Performing the cross-session
operations in the request handler has the consequence that neither
session type needs to keep references to the other.