kachel: support RECV_BUSY error on SEND/REPLY.
The error code RECV_BUSY indicates that the receiver is currently not
able to accept a message. In this case, the sender retries the command.
For that reason, the TCU state cannot change on errors, which required
some changes in the TCU. We also have new tests for these cases.