peidle: sync icache+dcache after application exit.
Apparently, the TCU always accesses the dcache when reading or writing
from/to memory. Thus, if we write new instructions to the dcache, we
have to make sure that the icache is in sync with the dcache before
trying to execute them. For that reason peidle now performs a fence.i
after the application exited.