resmng: support PE sharing with a resource manager.
There are some special things to take care of when the resource manager
shares its own PE with some child VPE. This is because in this case the
PMP EPs are already set by our loader. However, we do need to remember
the memory region to inherit these to later allocated PEs.