Only allow PMP configuration for root PE objects.
That is, all derived PE objects are not allowed to change PMP EPs. The
reasoning is that only applications that completely own a PE should be
able to configure PMP EPs. If an application is merely running on a PE
or is even sharing the PE with someone else, it should not be allowed
to change PMP EPs.

For that reason the resource manager passes a derived PE object to its
children unless they are a resource manager. In such a case, the
resource manager owns the PE and can configure PMP EPs. However, that
currently implies that resource managers cannot share their PE with
other VPEs on the same level. They can share their PE with child VPEs,
though. The reason is that the pager needs to configure a PMP EP for
the file system. It could be possible to let its parent resource
manager do that for him, but it's not worth the trouble I think.

If a new PE is allocated from a resource manager, the allocator owns
this PE and therefore receives a root PE object. As before, this PE
object has already the same PMP EPs configured as the allocator's PE.