kachel: deny READ/WRITE/SEND/REPLY if data contains a page boundary.
The change in M³ is simple for READ/WRITE, because the TCU abstraction
can simply take care that transfers don't cross page boundaries.
However, since the performance depends now on how the buffer is
aligned, I've added a AlignedBuf abstraction that contains a
page-aligned buffer of a specified size.

SEND/REPLY is more tricky, because we cannot split such a command into
multiple commands. For that reason, the message passed to the TCU
abstraction needs to be already placed in a way that it doesn't contain
a page boundary. To prevent that users accidentally forget that, I
introduced a new type called MsgBuf, which contains a 512-byte aligned
message of 512 byte. Note that this limits the message size to 512 byte
for now.