pemux: flush+invalidate cache on unmaps.
Since we have no cache coherence, we need to make sure that we load it
fresh from the DRAM the next time we use this memory. Thus, flush and
invalidate all cache lines on unmap. We might want to do that more
selectively later, though. However, at the moment unmaps almost never
happen during runtime, so that this not a big performance issue.