librustbase: added get_data_unchecked method to Message.
In Rust, data structures with large alignment requirements also have at
least this size. We need a 2k-aligned message buffer for the network
packets, but 2k is also the maximum message size of the TCU, including
header. Thus, the message payload is at most slightly less than 2k and
thus technically not long enough for the 2k-aligned message buffer. We
get around this problem by having a get_data_unchecked method that does
not check the size and alignment.