netrs: rewrote client API.
The new API is based on the NetEventChannel and DataQueue from Georg. I
adapted and simplified it for smoltcp and wrote the Rust version.

The current state has working UDP and TCP sockets, although more tests
are necessary to gain more confidence in the code.

For both C++ and Rust, we perform the minimum number of copies required
for the legacy-style API. That is, the packet data is copied once out
of the message in the NetEventChannel into the user buffer. We could
later also offer a different API that let's the user read directly from
the message in the NetEventChannel without performing a copy.