rust: introduce a default MsgBuf to reduce stack usage.
Allocating a new MsgBuf for every message becomes a problem in complex
applications, like root for example, because every MsgBuf consumes 512
byte stack space.

For that reason, I've now introduced a default MsgBuf instance stored in
BSS, which can be used for most purposes. We have to be careful though
to not use the default MsgBuf multiple times at once. Thus, I've
introduced MsgBufRef that keeps track of whether a buffer is in use or
not. For example, system calls cannot use the default MsgBuf, because
the application could use the default MsgBuf for a message over a
SendGate and the SendGate needs to be activated first via a syscall.

Furthermore, we need to be careful with our thread library to always
drop MsgBufRef before switching threads.

A nice side effect is that this also reduces cache and TLB pressure,
because we are now using the same memory most of the time for messages
leading to less cache/TLB misses in the core and also less TLB misses
in the TCU.