netrs: reworked server loop.
The server can sleep now if there is nothing to do. The basic idea is to
first push data to send into the smoltcp sockets, then poll smoltcp to
send and receive packets, and finally pull data out of the smoltcp
sockets. The poll function tells us whether something happened and thus
whether there might be new events to send to clients. The poll_delay
function tells us whether and if so how long we can sleep until the
next poll.

Additionally, we have to take the event channel into account. On the
receiving side (client -> server), we can only figure out after having
fetched an event message which socket is affected and thus whether the
socket has more send buffer space. If there is no space anymore, we
store the event for later processing. On the sending side
(server -> client), we now use credits in both directions and therefore
know in advance whether we have credits to send another event to the
client. If not, we don't look for state changes and don't pull data
from the smoltcp sockets.

That means, we cannot sleep in the server loop if there are events
pending for later processing. For the sending side, we get credits back
through messages and are thus waked up if that happens.