resmng: remember child waits and let operations fail on their removal.
If a child is removed, there could be an ongoing asynchronous operation
on behalf of that child. To make sure everything continues normally, we
now remember if that is the case and let this operation fail. This is
achieved by remembering the event the child is waiting for and waking
up the thread with a None-message, which ultimately lets the operation
fail.