netrs: introduced configurable quota for sockets and buffer space.
Clients can now specify the buffer sizes for each socket and have a
per-session quota, specified in the config files. Currently, you can
configure the max. number of sockets and the total available buffer
space. For stream sockets, each socket has a send and receive buffer.
For dgram sockets, there is an additional per-packet metadata, so that
the number of packets in that buffer can be configured, too.