netrs: let the server pick the local address.
The IP is defined by the server anyway and until we have multiple NICs,
there is no reason to pick anything different than this address. As
soon as we have multiple NICs, the client can probably specify the NIC
id or a wildcard to listen on all NICs or something like that.