kernel: notify PEMux even on explicit on RGate deactivation.
If there were still unread messages in the receive buffer, we have to
notify PEMux to adjust the total number of unread messages accordingly.
Forcing user applications to fetch theses messages before can be
difficult, because they have to do it before the deactivation
(afterwards the EP is invalid), but also need to make sure to not
receive further messages.

As this is not performance critical, let the kernel do that, since the
infrastructure is already there.