host: ensure that all processes are killed on errors.
It's a bit tricky to achieve that, because there does not seem to be a
tool that gives to-be-started command a new process group, and we start
the kernel indirectly via setarch/valgrind. Therefore, letting the
kernel set its own process group doesn't help either, because the
execute.sh does not know which one that is.

The simplest solution I could find is to write my own tool that uses a
new process group for the command to start. Then the execute.sh knows
that it's the pid of this tool and can simply kill that group after the
execution.