netrs: redesigned waiting concept.
The rational is as follows. The usage of one specific socket should only
care about this socket. That is, receiving should only process events
from this socket in order to receive data events and also to receive
state-change events. Sending should only fetch replies from this socket
to make sure that we get credits back and can send again. This behavior
should also be consistent between blocking and non-blocking mode.

For this to work, we cannot use VPE::sleep, but have to use
VPE::wait_for_msg in order to wait for messages on a specific RecvGate:
to receive events or receive replies, depending on whether we want to
receive or send. This way, we ignore messages on other sockets or any
other receive EPs we might have.

Additionally, applications occasionally want to wait until input/output
on _any_ socket is possible. This is particularly useful when working
in non-blocking mode. For that reason, the network manager provides the
methods wait and wait_for. These take a bitmask of directions
(input,output). The input side waits for data events and state-change
events, whereas the output side waits until new data can be sent
(credits are available).