net: reworked socket close/abort/drop.
To make it easy to use sockets correctly, sockets are now properly
closed in drop/destructor. That is, by default, TCP sockets go through
the close protocol, whereas UDP sockets make sure that all sent data is
indeed send.

For TCP, applications can opt out by calling abort before, which does a
hard and immediate abort of the connection at the cost of potentially
losing sent packets. Also, close can be called explicitly to properly
close the socket. With both close and abort, the socket can be reused
afterwards. For UDP, neither abort nor close exists, because there is
no close protocol to bypass and UDP sockets cannot be reused (e.g., if
a UDP socket is bound, there is no way to unbind it).

Furthermore, I moved all TCP-related code to the TCP module by making
the Socket members public. Socket is internal anyway and I think it is
better to have functions at the proper place than in Socket to keep the
members private.

Finally, I improved the non-blocking API to return false/errors in C++
and in general to work properly. There are also tests for the
non-blocking API now. In Rust the idea is: functions return WouldBlock
if the operation could not be started. Functions return InProgress if
the operation is started, but not finished yet. And functions return
AlreadyInProgress if the operation is started and the user tried to
start it again. C++ works analogously, but uses false/errors for
WouldBlock and InProgress and an exception for AlreadyInProgress.