net: introduce Endpoint type and added getters.
Due to the many places where both IpAddr and Port are needed, it's
better to have a type that contains both.

The TCPSocket provides the functions {local,remote}_endpoint now to let
applications know about their local IP/port and the remote IP/port. The
local endpoint is available if the socket is in listen mode or is
connected, whereas the remote endpoint is available if the socket is
connected (via connect or accept).

The UDPSocket provides only local_endpoint, which is available if the
socket was bound.