b: added workaround for broken build.ninja files.
If the user makes a mistake in the build.py's or similar, which results
in a broken build.ninja, the build.ninja was never regenerated so far.
This was because if ninja is given a broken build.ninja it stops
immediately and therefore does not detect that one of the dependencies
of the build.ninja has changed. So, if the user already fixed the bug
in the build.py, the build.ninja is still not regenerated.

To work around that problem, we now simply remove the "file id" file to
force a regeneration with the next b call. This is not optimal, because
we will also do that on compile failures, but since reconfiguration is
quick, I think it's acceptable.