Improved syntax of boot scripts.
The "sharepe" argument for root/pager was a weird special case that
specified that the first domain should be merged with the parent app's
domain. This argument has now been removed. Instead, it is now possible
to add "app"s as a direct child to an "app", i.e., without a "dom"
element. This can be used multiple times at any place.

Additionally, it is now mandatory to wrap the root app into a "dom"
element. Despite being removed before booting the system, I think this
rule makes it more logical, because every "dom" element in a boot
script corresponds to an isolated domain (own PE, own memory, ...) at
runtime. So, for example, you can simply count them to know the number
of PEs necessary for booting (ignoring the dynamically allocated PEs,
but these are also countable in the boot script).