hw: changed memory layout for virtual-memory support.
Therefore, we use the same memory layout and linkerscript as gem5 now.
Unfortunately, the environment location is a bit tricky. It should be
at a known location and of course cannot interfere with other things.
Placing it before code and data is difficult, because the entry point
on hw is always the start of DRAM. Placing it behind BSS is difficult,
because the kernel wants to extend the heap and the heap starts after
BSS and needs to be contiguous (and it still needs to be within the
first few MBs to fit inside the per-PE-region). Placing it within code
or data does not work, because somehow our loader in gem5 does run
before the executable is loaded (not sure why). In the end, I decided
to place it before the code by letting the code start at 0x10001000 and
writing a jump instruction manually to 0x10000000 (DRAM start), which
jumps to the actual code start. We then place the environment at
0x10000008.

Note however, that the environment location is different on other
targets and other ISAs.