hw: replaced print timeout with parameter-controlled timeout.
That is, we do no longer timeout after x seconds without print, but by
default we don't timeout at all and the user can use M3_HW_TIMEOUT to
let the execution timeout after the given number of seconds, regardless
of prints.

This approach is much more useful than the old one, because we can now
stop the execution via ^C anyway without losing the logs and when
running something interactively, a print timeout makes no sense. If we
run automated tests though, a timeout after a certain number of seconds
is useful, even when prints are happening, because that should indicate
that something went wrong.