pemux: improved mechanism to wait for IRQs.
Now we have a generic wait function that allows to wait for messages
and interrupts, both with an optional timeout. There are 4 cases:
1. ep=Some, irq=None: we wait until a message at given EP arrived
2. ep=None, irq=Some: we wait until the given IRQ arrived
3. ep=Some, irq=Some: we wait for either of these two
4. ep=None, irq=None: we wait for any message or any interrupt
Besides that, invalidated EPs and timeouts force unblock VPEs.

In contrast to before, interrupts need to be registered first. This is
currently done via PEXCall, but will later be done over the resource
manager and a sidecall.