Improved interrupt handling.
The potentially available PE-internal interrupts (except CORE_REQ and
TIMER) are now removed from the TCU-Interrupt struct. Instead, we just
specify them as numbers. I think in the end, we have to specify them in
the boot script and just pass them to the resource manager and then
finally via sidecall to PEMux. Potentially, we could introduce names
that are resolved by the platform description?

Additionally, PEMux has now a single interrupt handler for these cases
and we have an IRQSource enum that combines the TCU-related interrupts
and any "external" interrupts. This makes everything much clearer and
simpler.