resmng: introduced names for PEs.
Instead of letting the application specify the PE description when
allocating a new PE, we now let the application specify a name that
refers to the boot script. Thus, we can define the PE type to use in
the boot script.

Additionally, we now use attributes to represent special features of a
PE like an attached NIC or a specific ISA implementation like Rocket or
BOOM. This allows us to let run any, e.g, RISC-V binary on such a PE
and still differentiate between the PEs.