rust: Make MemGate::activate() and MemGate::ep() public
This can be used by applications/services to implement functionality
similar to the file protocol where one VPE delegates an EP capability
to another which then configures the memory endpoint.

MemGate::ep() is needed to revoke the EP capability again later
(also similar to how it is already done in the GenericFile implementation).