rust: vfs: Store which EP selector was delegated
Prepare for temporarily delegating another EP in case parts of the
file should be read by some remote service (e.g. the hash multiplexer).

The general idea is that all the typical file operations can be used
(e.g. seek) but then parts of the file can be directly read by the
hash multiplexer without having to copy them to the current VPE.