rust: vfs: Forcefully commit when switching delegated EP
When delegating a new EP the file server will not immediately
configure the new EP for the currently visible part of the file,
this happens only on the next NEXT_IN/NEXT_OUT operation. Therefore,
it's important that this happens when switching the delegated EP.

To ensure this, commit the bytes that were already read/written
before delegating the new EP. This means that the remaining bytes
can be requested again using a new NEXT_IN/NEXT_OUT.

Also, avoid calling NEXT_IN/NEXT_OUT if len == 0 (nothing should be
read/written for some reason?). This would break the logic because
0 bytes cannot be committed in the protocol. In that case, there
would be no way to configure the new EP with the same memory again.