m3fs: Fix file offset when committing after replacing some bytes
At the moment, replacing some text in an existing file (e.g. writing
"Hello " to a file with content "Hallo World!") and then committing
the result (e.g. by reading the rest of the file) will produce an
incorrect file offset. M3FS assumes the previous extent was fully
read/written even though only some of the bytes were replaced.

This seems to happen because of an inverted comparison in commit()
in M3FS. The current extent should be kept if off + nbytes is still
within the current extent, not the other way around.

Add a test to reproduce the problem and fix the if condition to
make the test pass.