rustunittests: thash: Add unit tests for hash multiplexer (v2)
Try to calculate various hashes using the hash multiplexer to validate
the context switching as well as the results from the hash accelerator.

The "hex-literal" crate is used to create easily-readable byte slices
for the resulting hashes (determined e.g. with sha3sum for comparison).

v2: Keep file open that is passed to children VPEs. M3FS closes
    file session recursively now.