rustunittests: thash: Add test that hashes mapped memory
The memory shared with the hash multiplexer must be page-aligned and
physically contigouous. The best way to ensure that seems to be to
create a new MemGate and then map it with map_mem(). After writing
to the mapped memory it is important to flush the cache (or at least
the region of the buffer), otherwise the changes might not be seen
by the hash multiplexer.