hw: added support for serial input.
The current approach uses a polling-based interface between the Python
script and a PEXCall on the FPGA and is thus suboptimal. However, it's
sufficient for our purposes at the moment. We can later replace it with
messages sent by the Python script to a specific EP on a specific PE.