libbase: catch exceptions from main and lambdas.
Apparently, the default terminate handler does not destruct objects on
the unwinded stack leading to problems. For example, we could omit a
TCU ack and fail later with insufficient space in the receive buffer.
Therefore, catch all otherwise uncaught exceptions in env_run to ensure
that everything is properly cleaned up before continuing.

This requires to enable exceptions for libbase, but this seems to be
okay now as most stuff is written in Rust anyway.