Introduced a new approach to PE selection.
The previous approach of naming PE resources assigned to applications in
the config file had several disadvantages. At first, it made it
difficult to use applications in multiple configs as they had to agree
on these names. Second, it required the user to know what names
applications are using. Third, it required hacks like "child0" to
select the PE for a specific child. Finally, applications could not
communicate what they actually need. For example, an application might
require that two PEs are identical, but the config allows to choose any
PE without restrictions.

Due to all of these problems, I completely redesigned the PE selection.
The idea is that applications use an API that selects a PE based on a
textual description and it tries to find the best PE available for this
application. The textual description consists of multiple words
separated by '|' that are tried in the given order. Each of these words
describes a PE that would match the requirements and in turn consists
of multiple '+' separated attributes (e.g., 'boom+nic'). Additionally,
'own' can be used to denote the own PE (requiring PE-sharing support)
and 'clone' can be used to denote a separate, but identical PE.

For example, an application could say PE::get("own|boom|core"). The get
method would then try to find a matching PE, starting with 'own'. If
the own PE supports sharing, the method simply returns another
reference to the own PE. If not, get tries to allocate a new PE with
the description 'boom'. If the resources assigned to this application
still contain a BOOM core, this one is allocated. Otherwise, get tries
to allocate any core.

In the config files, the 'pes' element has no longer a name, but just
specifies the type of PE using the same notation as above except
that 'clone' and 'own' are not supported. So, it can also be something
like 'boom|core' to prefer a BOOM core, but take any core in case a
BOOM is not available. To select the PE used by the resource manager
for a specific domain, the 'pe' attribute of the domain element is
used, which supports the same notation as in the 'pes' element. That is,
to put an application preferably on a Rocket, you specify:
<dom pe="rocket"><app ... /></dom>
Note that a rocket has to be available in case the resource manager is
not root. So, you might need to add a 'pes' element to the resource
manager's resources.