libm3: use strtok for PE::get.
Apparently, there is no sensible way to split a string into tokens in
C++ < 20. We can use stringstream and getline, which can directly work
on a read-only string, but unfortunately this needs heap allocations,
string copies and most importantly increases the size of our binaries
by 500K (!!!). Alternatively, we can use strtok, but this changes the
string so that we need to make a copy first. However, it still seems to
be the better choice here.