Reworked input concept.
So far, we had a pretty simple way to handle input on host and gem5 and
only a preliminary polling-based approach within PEMux for input on the
FPGA. In all cases we just read from stdin in cooked mode and let Linux
do the line editing.

Working towards proper input processing, I now reworked the whole
concept. I thought a lot about whether to just use a UART or whether we
want a different solution. In the end, I decided against UART due to
our special requirements. Since we have a tiled architecture and M³
behaves as a distributed system, using a UART would force us to connect
it to one specific PE. For that reason, debug prints need quite some
infrastructure to work, because they have to be send to another PE. To
minimize that infrastructure and thereby lower the barrier for
debugging, we want to continue to use the TCU print for debug prints. I
considered to use a UART for all "real" prints and for input, but this
leads to the problem that we have two independent channels for prints.
Thus, prints are no longer properly ordered.

Due to these reasons, I decided that we want to have a custom way for
input and output. Output is sent via a message to the Ethernet IF,
whereas input is received as a message from the Ethernet IF. Output can
be done by all components via the TCU print operation, whereas
receiving input requires to configure a receive EP first.

To implement this simple concept, the kernel creates a special receive
gate and passes it to root. I extended the resource manager by
a "get_serial" operation that delegates this receive gate to the
caller. Using this operation requires the <serial /> element for the
calling application in its config. Currently, only vterm is using that
and therefore needs this element. The shell and applications started by
it get connected to the vterm and can therefore do input and output via
vterm. Currently, vterm always works in raw mode and therefore passes
raw input bytes to the caller. The shell supports some basic line
editing based on that.