host: make TCU communication blocking in Rust.
That is, we do no longer poll for the CPU-TCU communication and also do
no longer poll for the TCU-TCU communication. The TCU thread uses
select to wait until a message arrived. The CPU thread uses a blocking
receive to wait for command completions. Additionally, the TCU supports
a SLEEP command, analogous to the gem5 TCU.