host: support sleep with timeout and custom fds to wait for in Rust.
For networking, we need to be able to sleep until either a message
arrives or a timeout is reached. Additionally, the host NIC driver uses
a socket to the kernel to send or receive packets. In order to wake up
if packets have been received, it tells the TCU additional file
descriptors to pass to select.