kernel: ensure that every VPE can occupy at most one thread.
The vpe_wait system call allows to wait synchronously (the kernel does
not reply until one of the VPEs exited) and asynchronously (the kernel
replies immediately and sends an upcall as soon as a VPE exited). Both
cases were implemented by letting the thread wait until any VPE exited.
Since the async version allowed to perform other syscalls in the
meantime, the VPE could require one more thread in case another
blocking system call is used.

To ensure that every VPE can occupy at most one thread, I changed the
implementation of the async version by remembering a list of
VPEs-to-be-notified globally and just send upcalls on exits. That is,
for the async version, no thread needs to wait, but the thread that
handles a VPE-exit will just send all upcalls as required.