fs: move servers into sbin and exclude non-supported binaries.
The folder /bin should only contain applications. Not all of them can be
executed from the shell due to missing permissions or similar, but at
least in theory they can. Servers should go to /sbin as it makes no
sense to execute them from the shell. Other binaries such as the
kernel, PEMux, or baremetal/standalone programs should not be put into
the FS at all.