vterm,shell: terminate child VPEs on ^C.
Applications that have a channel at vterm can now delegate a signal
SendGate to vterm, which vterm uses to send a message upon ^C to the
application. Currently, all clients that delegate a SendGate get the
signal, but only the shell uses it. We can extend it later to only send
it to the foreground client.