resmng: make sharing policies more consistent.
We now have a clear policy of how resources are shared if nothing
specific is requested. By default, resources (EPs, kernel mem, and user
mem) including the quota are shared among all VPEs on a PE. If a
specific amount is requested for a VPE, this VPE receives an exclusive
quota. Resources are never shared between PEs.

The rational behind this is that the default case should just work and
create as few problems as possible. Thus, we just share a resource so
that the portion of each app is determined dynamically. If an app has
specific requirements (e.g., needs at most N EPs or M bytes of user
memory), then this can be specified in the config file and the app will
receive a separate and isolated share of that resource. In other words,
this app is guaranteed to receive the requested share of that resource.