rust: Allow using LazyReadOnlyCell for RequestHandler
At the moment the RequestHandler cannot be placed in a LazyReadOnlyCell
because the handle() function requires a mutable self reference. However,
there is actually no need for this, it just calls RecvGate::fetch()
which can be used read-only.

Drop the mut from the self parameter to allow using LazyReadOnlyCell for
the RequestHandler. This avoids problems at runtime when the RecvGate in the
RequestHandler is needed while handling a request. It is also a bit faster
because there is no need for borrow checking at runtime.