m3fs: store meta buffer in unsafe global variable.
Unfortunately, it's really hard to use a safe cell for the meta buffer,
because we often need mutable access to multiple (different) places of
the meta buffer at the same time. Additionally, these accesses are
spread all over the codebase and are sometimes nested.

We therefore continue to use an unsafe cell for it.