Configure pager EPs from via separate syscalls.
Previously, the create_vpe syscall expected a send and receive gate and
activated these manually on the dedicated pager EPs. This made the
create_vpe syscall more complicated than necessary and actually, this
can also be done from applications.

Therefore, now the VPE that creates a child VPE is responsible to
activate the pager gates accordingly. That works mostly very smoothly,
but unfortunately requires a special code path in the kernel to
properly configure the receive EP. The reason is that applications
don't have access to the standard-receive-buffer memory and thus can't
specify the capability for that memory. Allocating separate memory
would be possible, but gets pretty complicated, because PEMux also
needs to know that address. Therefore, I decided to keep it simple for
now and introduce a special case for the dedicated pager receive-EP and
just calculate the address, as before, in the kernel.