Added support for file-session multiplexing.
Most of the time, applications don't want to delegate files to others,
but just want to use them for themselves. Therefore, all the capability
exchanges and revokes during open and close are actually not
necessary.

For that reason I added file-session multiplexing that enables
applications to open and close files within their meta session.
Additionally, endpoints to get access to the data are reused to avoid
even more useless delegations and revocations. This approach does not
only improve performance but also scalability, because we don't involve
the kernel for open and close. Therefore, open-close-heavy benchmarks
like sqlite perform and scale significantly better.

Note that this is a cleaner version of this feature. The previous
version was much more limited and hacky, which is why I ultimately
removed it from the codebase. The new version is transparent for
applications (does not need to be setup beforehand), supports an
arbitrary number of cached endpoints and is integrated more cleanly.

Due to the transparency and better integration, files are also opened
within the meta session by default. Applications that want to delegate
a file to somebody else therefore need to request a new file session
for that file via the OpenFlags::NEW_SESS flag. Currently, this is only
used by the shell and when mapping files like in VPE::exec.