libm3: removed now obsolete Gate-list.
A single gate can only be activated once. For that reason, already
activate gates should not be delegated, because the receiver would not
be able to activate it again. Instead, the gate should be deactivated
before delegating it.

This implies that the case the list tried to address does not exist
anymore. Either the gate was not activated, then it would not be in the
list anyway, or it had been activated, but then we can't activate it
again even if we pretend we didn't do so on the user side.