Reworked VPE::run.
The ability to clone the current program and then execute a given
function caused lots of trouble since years and therefore has made me
think about alternatives several times. The underlying problem is a
mental split between two worlds:
1. the capability world
2. the UNIX fork&exec word

I always liked aspects of both and therefore tried to have some of both
in M³. However, after thinking about it again and again, I'm now
convinced that having both is not achievable. The abstract reason is
that the capability world prefers deny-by-default, whereas the UNIX
world prefers allow-by-default. Therefore, the clone feature of VPE
attempts to get the best of both: giving users the convenience of fork,
but starting with a fresh set of capabilities.

This primary reason why I always wanted this feature (and still do!) are
unittests. Unittests frequently need tests that involve multiple VPEs
to, for example, let one VPE read from a pipe and another VPE write to
the pipe. However, creating completely separate programs for both would
be *really* annoying, because you couldn't easily transfer values from
one to the other, would have them in separate files, and would create
an insane amount of binaries just for tests. Therefore, just cloning
yourself and letting the clone call a lambda is exactly what we want.

However, after multiple years of pain regarding this issue, I have to
conclude that this clone feature is simply not supportable if the clone
needs to be partial (to start with fresh caps, endpoints, etc.). To
illustrate that, let's pick one of the many problems. Let's say, VPE1
creates a SendGate S and creates a clone of itself (VPE2). The S object
lives somewhere in VPE1's address space and will be automatically
activated upon usage, i.e., an EP will be allocated and configured
accordingly. Now VPE2 has the same S object. Since all objects in M³
use RAII, both VPE1 and VPE2 will deconstruct the object and thereby
revoke the cap (bad!). Furthermore, if VPE1 already used S, it is
activated. If VPE2 tries to use S as well, the object thinks its
activated, but the EP is invalid in VPE2's TCU. Worse, if VPE2
activates another gate (S2), it can happen that S2 receives the same EP
id as S, so that operations with S work afterwards, but use the wrong
EP. Note also that S cannot be activated again, because that would lead
to two EPs and thereby twice as many credits, effectively defeating the
credit system. There are various other subtle problems with global
objects, or application-internal caches (i.e., many objects need to
be "reset" to reflect that caps/EPs are not there, which led to bugs
several times).

Therefore, I now decided for a different approach: we support VPE::run,
but without cloning the program. Instead, we start the same program
again (fresh state) and offer communication channels (as before) to
transfer data to the child. I think, this reflects exactly the
semantics that we have: you start with a fresh state and everything
that you need has to be explicitly passed to the new VPE. As before,
files and mount points can be passed to the child VPE. Furthermore, plain
data can be transferred to the child.

This has a couple of limitations, though. At first, we require that the
program is available in the file system and we have access to it with
the same path as we were started. And second, we cannot use capturing
closures anymore, because there does not seem to be a way to copy them
to the child (integers are sometimes [architecture/language dependent]
not copied into the closure, apparently, and objects can only be
shallow-copied). Considering that VPE::run is basically exclusively
used for tests, I think we can live with these limitations.