rust: improved handling of uninitialized memory.
Uninitialized memory in unsafe code is actually pretty complicated and
still under discussion within the Rust community. A good article about
it is: https://www.ralfj.de/blog/2019/07/14/uninit.html

One important insight is that the compiler is not just free to assume
any particular value for uninitialized memory, but can even chose
a *different* value for every read operation. And this is not just
theory, but the compiler actually takes advantage of that for
optimizations :)

For that reason, it is not okay to read uninitialized memory, even if
these are integers where every bit pattern is valid. There were a
couple of places that could lead to undefined behavior if a called
function (or similar) accessed the uninitialized memory.

In places that don't matter much for performance, I now went the easy
way: just initialize the memory. This is unfortunate at some places
(e.g., Read::read_string), but difficult to fix, because we call an
unknown function with a pointer to that memory.

There are other places, that are very performance critical, such as
MemGate::read_into_vec, that still use the Vec::with_capacity and
Vec::set_len trick, but we know that the memory is not read before it
is initialized (e.g., TCU::read is used to initialize it).

Finally, some places already used MaybeUninit, but in a bad way. For
example, MsgBuf now actually stores an array of MaybeUninit<u8> and
transmutes it to an u8 slice at places where we are sure that this
slice is initialized.