kernel: extend the heap eagerly.
The problem with a dynamic heap extension is that we cannot allocate any
memory during the extension. For example, the Error type in debug mode
contains a Box with more detailed information about the error
(backtrace etc.). Therefore, as soon as an error happens, we allocate.
And in fact, an "error" is happening in case the TCU TLB does not
contain an entry for a virtual address we invalidate. Note that the
allocation causes a panic due to the StaticRefCell that is already
borrowed for the current heap allocation that required the heap
extension and will panic if the heap extension tries to allocate again.
But even if we resolve this problem, we have the general problem that
there simply might not be any memory on the heap left (we extend it
after all).

Without measures like knowing how much we might allocate during
extension in the worst case and keeping that memory around or similar,
there is no way to dynamically extend the heap, I'm afraid. However,
the question is whether dynamically extending the heap is worthwhile at
all. Since we know the available memory upfront and don't need this
type of memory (KERNEL) for anything else but the heap after
initialization, there is actually no good reason to allocate it lazily.
At least provided that the allocation/mapping of the memory doesn't
take too long. Therefore, we allocate everything eagerly now and use
large pages to prevent that we need to create tens of thousands of
PTEs.